modifier edit category

+++ src/Controller/Admin/CategoryCrudController.php
@@ -120,22 +120,23 @@ class CategoryCrudController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
             $logo = $form->get('pathLogo')->getData();
+            if($logo != null){
+                //on genere un nouveau nom de fichier (codé) et on rajoute son extension
+                $fichier = md5(uniqid()) . '.' . $logo->guessExtension();
 
-            //on genere un nouveau nom de fichier (codé) et on rajoute son extension
-            $fichier = md5(uniqid()) . '.' . $logo->guessExtension();
+                // on copie le fichier dans le dossier uploads
+                // 2 params (destination, fichier)
+                $logo->move($this->getParameter('logo_directory'), $fichier);
 
-            // on copie le fichier dans le dossier uploads
-            // 2 params (destination, fichier)
-            $logo->move( $this->getParameter('logo_directory'),$fichier );
+                /* Penser a supprimer les ancien fichiers  */
+                unlink($this->getParameter('logo_directory') . '/' . $category->getPathLogo()); //ici je supprime le fichier
 
-            /* Penser a supprimer les ancien fichiers  */
-                
-            unlink($this->getParameter('logo_directory').'/'. $category->getPathLogo()); //ici je supprime le fichier
-               
-
-            // on stock l'image dans la bdd (son nom)
-            $category->setPathLogo($fichier);
+                // on stock l'image dans la bdd (son nom)
+                $category->setPathLogo($fichier);
+            }
+            
 
             $category->setName($request->request->get('category')['name'])
                 ->setSlug(strtolower($this->slugger->slug($category->getName())))
